move v2tov3 converter to storage-v2-converter
