+++ src/infra/chromeStorageClient.ts
@@ -1,15 +1,6 @@
-import { v4 as uuidv4 } from "uuid"
 import { IChromeStorageClient } from "./interface/chromeStorageClient"
-import {
-  Storage_v3,
-  KEY_VERSION_2,
-  KEY_VERSION_3,
-  ItemValue,
-  BlockValue,
-} from "./scheme"
+import { Storage_v3, KEY_VERSION_2, KEY_VERSION_3 } from "./scheme"
 import { Storage_v2 } from "./scheme-v2-client/interface/scheme"
-import { IShopItem } from "../shopItem"
-import { stringToNumber } from "../utils/stringToNumber"
 
 export class ChromeStorageClient implements IChromeStorageClient {
   async storageV2LocalGet(): Promise<undefined | Storage_v2> {
@@ -45,87 +36,5 @@ export class ChromeStorageClient implements IChromeStorageClient {
         }
       )
     })
-  }
-
-  convertStorageV2ToV3(storage_v2: Storage_v2): Storage_v3 {
-    const itemsConverted = Object.fromEntries(
-      storage_v2.shopItems.map((item, index) => {
-        const id = uuidv4()
-        return [id, ChromeStorageClient.convertV2toV3(id, index, item)]
-      })
-    )
-    const keys = Object.keys(itemsConverted)
-    const firstItemId = keys.length ? keys[0] : null
-
-    return {
-      selectedItemId: firstItemId,
-      items: itemsConverted,
-    }
-  }
-
-  private static convertV2toV3(
-    id: string,
-    index: number,
-    item: IShopItem
-  ): ItemValue {
-    const { descriptions, details, imageURL } = item
-    const imageBlock: BlockValue[] = imageURL
-      ? [
-          {
-            id: uuidv4(),
-            type: "image",
-            value: {
-              imageUrl: imageURL,
-            },
-          },
-        ]
-      : []
-    const textBlocks: BlockValue[] = descriptions.flatMap<BlockValue>(desc => {
-      return [
-        {
-          id: uuidv4(),
-          type: "heading",
-          value: {
-            content: desc.title,
-          },
-        },
-        {
-          id: uuidv4(),
-          type: "text",
-          value: {
-            content: desc.body,
-          },
-        },
-      ]
-    })
-    const tableBlock: BlockValue[] = details.length
-      ? [
-          {
-            id: uuidv4(),
-            type: "table",
-            value: {
-              rows: details.map(detail => {
-                return {
-                  title: detail.title,
-                  body: detail.body,
-                }
-              }),
-            },
-          },
-        ]
-      : []
-    const blocks: BlockValue[] = [...imageBlock, ...textBlocks, ...tableBlock]
-
-    return {
-      id,
-      name: item.name,
-      price: stringToNumber(item.price),
-      weight: stringToNumber(item.weight),
-      stockRakuten: stringToNumber(item.stockRakuten),
-      stockMakeshop: stringToNumber(item.stockMakeshop),
-      jancodeString: item.jancode,
-      blocks,
-      order: index,
-    }
   }
 }
